Extract both scoping versions
