+++ internal/commands/scope_selection_display.go
